refactor: improve merge commit detection logic
Enhanced the isMergeCommit function to better distinguish between
different types of commits by comparing tree IDs and checking parent
commits more accurately. This change improves the reliability of the
commit-msg hook when determining whether a commit is a merge commit.

- Use git rev-parse HEAD^{tree} to get the HEAD commit tree ID
- Compare staged tree ID with HEAD tree ID to detect new changes
- Check parent commits of HEAD to determine if it's a merge commit
- Improve comments and logic flow for better clarity

This addresses potential issues with the previous implementation that
only checked if the message file was MERGE_MSG or if the first line
started with "Merge ", which could be less reliable in certain scenarios.

Change-Id: Ic291e0e96a4d32821a53f5278d13b904215471fc
Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>